Consolodate v1 integration tests
